PyBulletHell/pybullethell/requirements.txt
pygame-2.1.0 (SDL 2.0.16, Python 3.10.1)